--- src/main/java/gift/exception/product/ProductExceptionHandler.java
@@ -1,20 +1,0 @@
-package gift.exception.product;
-
-import gift.controller.ProductController;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice(assignableTypes = ProductController.class)
-public class ProductExceptionHandler {
-    @ExceptionHandler(value = MethodArgumentNotValidException.class)
-    public ResponseEntity<String> handleProductNameException(MethodArgumentNotValidException e) {
-        String errorMessage = e.getBindingResult().getFieldErrors().stream()
-                .map(error -> error.getDefaultMessage())
-                .findFirst()
-                .orElse("올바르지 않은 입력 방식입니다.");
-        return ResponseEntity.badRequest().body(errorMessage);
-    }
-}
