--- src/main/java/gift/exception/wish/WishExceptionHandler.java
@@ -1,19 +1,0 @@
-package gift.exception.wish;
-
-import gift.controller.WishController;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice(assignableTypes = WishController.class)
-public class WishExceptionHandler {
-    @ExceptionHandler(value = MethodArgumentNotValidException.class)
-    public ResponseEntity<String> handleWishRequestValidException(MethodArgumentNotValidException e) {
-        String errorMessage = e.getBindingResult().getFieldErrors().stream()
-                .map(error -> error.getDefaultMessage())
-                .findFirst()
-                .orElse("올바르지 않은 입력 방식입니다.");
-        return ResponseEntity.badRequest().body(errorMessage);
-    }
-}
